requirements.txt
python-telegram-bot==21.10
openai==1.61.0
python-dotenv==1.0.1
gitpython==3.1.44
pydantic==2.10.6
